parser: Split code into multiple files Ergonomic 👍
