+++ parser/err.go
@@ -27,3 +27,19 @@ func (e *ParserError) Print() {
 	fmt.Println("   ", e.Where)
 	fmt.Println()
 }
+
+func (p *Parser) selfError(where *lexer.Token, msg string) error {
+	return &ParserError{
+		msg:   msg,
+		cause: nil,
+		Where: where,
+	}
+}
+
+func (p *Parser) otherError(where *lexer.Token, msg string, cause error) error {
+	return &ParserError{
+		msg:   msg + ": " + cause.Error(),
+		cause: cause,
+		Where: where,
+	}
+}
